+++ Sockets/Sockets.Implementation.WinRT/UdpSocketReceiver.cs
@@ -14,16 +14,27 @@ namespace Sockets.Plugin
         /// <summary>
         ///     Binds the <code>UdpSocketReceiver</code> to the specified port on all endpoints and listens for UDP traffic.
         /// </summary>
-        /// <param name="port">The port to listen on.</param>
+        /// <param name="port">The port to listen on.</param>        
+        /// <param name="listenOn">The <code>CommsInterface</code> to listen on. If unspecified, all interfaces will be bound.</param>
         /// <returns></returns>
-        public async Task StartListeningAsync(int port)
+        public async Task StartListeningAsync(int port, ICommsInterface listenOn = null)
         {
+            if (listenOn != null && !listenOn.IsUsable)
+                throw new InvalidOperationException("Cannot listen on an unusable interface. Check the IsUsable property before attemping to bind.");
+            
             var sn = port.ToString();
-
-            await _backingDatagramSocket.BindServiceNameAsync(sn);
+#if !WP80    
+            if (listenOn != null)
+            {
+                var adapter = ((CommsInterface) listenOn).NativeNetworkAdapter;
+                await _backingDatagramSocket.BindServiceNameAsync(sn, adapter);
+            }
+            else
+#endif
+                await _backingDatagramSocket.BindServiceNameAsync(sn);
         }
 
-        /// <summary>
+        /// <summary>   
         ///     Unbinds a bound <code>UdpSocketReceiver</code>. Should not be called if the <code>UdpSocketReceiver</code> has not
         ///     yet been unbound.
         /// </summary>