+++ Sockets/Sockets.Plugin/UdpSocketReceiver.cs
@@ -13,8 +13,9 @@ namespace Sockets.Plugin
         ///     Binds the <code>UdpSocketReceiver</code> to the specified port on all endpoints and listens for UDP traffic.
         /// </summary>
         /// <param name="port">The port to listen on.</param>
+        /// <param name="listenOn">The <code>CommsInterface</code> to listen on. If unspecified, all interfaces will be bound.</param>
         /// <returns></returns>
-        public Task StartListeningAsync(int port)
+        public Task StartListeningAsync(int port, ICommsInterface listenOn = null)
         {
             throw new NotImplementedException(PCL.BaitWithoutSwitchMessage);
         }